Create ~/.gemrc before updating rubygems
Previously, the only gem referred to in the build was Bundler.
Immediately before it was installed, rubygems was globally configured to
not install documentation when gems are installed or updated.

Now, rubygems is updated in the same RUN step that installs ruby.  By
creating the .gemrc file before that step, the `gem update rubygems`
command no longer needs an explicit `--no-documentation` flag.