Replace BUNDLE_BIN with BUNDLE_BIN_DIR
By specifying BUNDLE_BIN, bundler will generate binstubs for every gem, which
results in it generating a binstub for itself. The binstub always points to the
same Gemfile, so it is no longer possible to use the bundler on other Gemfiles.

This rename is just to prevent this conflict from occurring.

More info:

[BUNDLE_BIN](https://bundler.io/v1.16/bundle_config.html#LIST-OF-AVAILABLE-KEYS),